requirements.txt
aiogram>=3.0.0
telethon>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0